+++ edzed/blocklib/cron.py
@@ -14,10 +14,10 @@ import bisect
 from dataclasses import dataclass
 import time
 
-from . import addons
-from . import block
-from . import utils
-from .utils import timeinterval
+from .. import addons
+from .. import block
+from .. import utils
+from . import timeinterval
 
 # HMS is an abbreviation for clock time hour, minute, second
 # MD is an abbreviation for month, day